+++ src/components/animated/ChatFlowContent.jsx
@@ -164,7 +164,7 @@ export default function ChatFlowContent() {
       <NotificationsWrapper>
         <NotificationOne>
           <MessageNotification
-            link="entrepreneurs"
+            link="enterprise"
             name="Julius"
             text="Hey, I'm about to rebrand my company website, can you help me create it? 👋🏼"
             image="images/profiles/julius.png"
