+++ src/components/layout/navigation/menuTooltip.jsx
@@ -1,0 +1,59 @@
+import React from "react"
+import styled from "styled-components"
+import { tooltipData } from "./menuData"
+import MenuButton from "./MenuButton"
+
+export default function MenuTooltip(props) {
+  const { isOpen } = props
+  return (
+    <Wrapper isOpen={isOpen}>
+      {tooltipData.map((item, index) => (
+        <MenuButton item={item} key={index} />
+      ))}
+    </Wrapper>
+  )
+}
+
+const Wrapper = styled.div`
+ padding: 20px;
+ position: absolute;
+ top: 70px;
+ right: 30px;
+ opacity: ${props => (props.isOpen ? 1 : 0)};
+ z-index: 1;
+ background: rgba( 255, 255, 255, 0.30 );
+ box-shadow: 0 8px 32px 0 rgba( 31, 38, 135, 0.37 );
+ backdrop-filter: blur( 10.0px );
+ -webkit-backdrop-filter: blur( 10.0px );
+ border: 1px solid rgba( 255, 255, 255, 0.18 );
+ border-radius: 20px;
+ display: grid;
+ gap: 10px;
+ grid-template-columns: 150px;
+ transition: 0.3s ease-in-out;
+ /* display: ${props => (props.isOpen ? "block" : "none")}; */
+ visibility: ${props => (props.isOpen ? "visible" : "hidden")};
+ transform: ${props =>
+   props.isOpen
+     ? "skewY(0deg) rotate(0deg) translateY(0px)"
+     : "skewY(-5deg) rotate(5deg) translateY(-30px)"};
+
+     border-radius: 1.25rem;
+     box-shadow: 0 8px 32px 0 rgba(0, 0, 0, 0.37);
+   
+     @supports (backdrop-filter: blur(40px)) or
+       (-webkit-backdrop-filter: blur(40px)) {
+       opacity: 0, 5;
+       background-color: transparent;
+       -webkit-backdrop-filter: blur(40px);
+       backdrop-filter: blur(40px);
+       border: 1px solid rgba(255, 255, 255, 0.18);
+     }
+   
+     @-moz-document url-prefix() {
+       background-color: rgba(242, 242, 242, 0.7);
+       @media (prefers-color-scheme: dark) {
+         background-color: rgba(50, 50, 52, 0.7);
+       }
+     }
+`
